Move more build logic into Toothpick
