+++ buildSrc/build.gradle.kts
@@ -1,4 +1,5 @@
 val kotlinxDomVersion = "0.0.10"
+val shadowVersion = "6.1.0"
 
 plugins {
     `kotlin-dsl`
@@ -7,10 +8,12 @@ plugins {
 repositories {
     mavenCentral()
     jcenter()
+    maven("https://plugins.gradle.org/m2/")
 }
 
 dependencies {
     implementation("org.jetbrains.kotlinx:kotlinx.dom:$kotlinxDomVersion")
+    implementation("com.github.jengelman.gradle.plugins:shadow:$shadowVersion")
 }
 
 gradlePlugin {
